+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## V1.1.6
+
+setList use resetData(which call sortList#replaceAll)  
+add rvCompare rvAreContentsTheSame rvAreItemsTheSame method to quick access sortList
+
 ## V1.1.5
 
 fix IndexOutOfBoundsException when get -1 item